REFACTOR-#6046: remove code duplication for 'progress_bar_wrapper'
